+++ .vscode/settings.json
@@ -114,5 +114,22 @@
 		"transition",
 		"width",
 		"zIndex"
+	],
+	"cSpell.words": [
+		"Bernense",
+		"fullpage",
+		"grouphover",
+		"iiif",
+		"itemstructure",
+		"Katja",
+		"licence",
+		"openseadragon",
+		"Rembold",
+		"returnitems",
+		"rotateleft",
+		"rotateright",
+		"Seadragon",
+		"zoomin",
+		"zoomout"
 	]
 }
