+++ src/app.postcss
@@ -3,6 +3,11 @@
 @tailwind utilities;
 @tailwind variants;
 
+html,
+body {
+	@apply h-full overflow-hidden;
+}
+
 @font-face {
 	font-family: 'Inter';
 	src: url('/fonts/Inter-VariableFont_slnt,wght.ttf');
